Use done file as cache for the Director file
Director can recreate a jail if the user modifies the Director file
or there is another change that it tracks (such as a Makejail),
which means that the jail is destroyed and recreated. To avoid this
due to user error, the done file (which is currently used to indicate
that the VM has been successfully created) is used to save the
contents of the Director file.