Generate a random secret key by default
By default, a secret key from the configuration file is used.
However, if the user does not change it this will insecure his instance.
Normally a user who knows what they are doing should change the secret key
to generate tokens, however to be a little more secure by default a
random secret key is generated when the user does not specify one in the
configuration file.