Add `script-environment` parameter
It is very useful to provide environment variables along with metadata,
so that a user can further configure a virtual machine. Metadata provides
configuration to many virtual machines, but environment variables to only
one. For example, in a metadata I can specify an `authorized_keys` file
that many VMs can use, and that is installed using the script, however,
it is better to specify a the value of a parameter such as `hostname`
in an environment variable.