Add `appConfig` as new deployment type
A new deployment is here, its name is `appConfig` and the intention
is to create more templates for deployments, improving usability
when deploying things in many chains. You deploy a metadata that
includes the Mako template and that will be used in an `appConfig`
deployment, its variables are replaced by the values you have defined,
and you have a new deployment. In this effect, `appConfig` deployments
are like deploying applications as configuration files.