Add `autodisable.strict` parameter
Before this change, clients can connect without any problem to blacklisted
chains (due to smart timeouts), however after testing failures it is better
to be strict and disable a chain completely not allowing to connect after
the smart timeout ends.

The test was to make 100 HTTP requests with and without being strict. Being
strict, the chain was disabled after detecting some failures in one chain,
and prevented more than 80% of HTTP requests (wasting resources), this is
due to a combination of the retry policy and a good configuration on
timeouts (clients and chains), so the remaining clients waiting for HTTP
response will benefit after a while.