Execute labels when forcibly destroy the projects
To improve consistency when forcibly destroying a project, labels
are now executed. Prior to this change the labels were not executed
(while forcibly destroying a project) and the project was destroyed
anyway, however this is problematic as services using integrations
may have data referencing things that do not exist (e.g. pointing
to a non-existent service in SkyDNS). In the event that any
integration fails, the project is not destroyed and the administrator
must fix the problem.