Set the HTTP status code of the chain
By default, 503 is used as the status code when an exception occurs
when executing a remote function from one chain to another.
The problem with this approach is that it will trigger the retry
policy, which is undesirable in some cases. For example, if a user
requests a log from a chain of length 3 (e.g. `A.B.C`) but that
log cannot be found, the last chain (i.e. `C`) will return 404
but `B` will return 503 which triggers a retry policy from `A`.