Display status even when the project does not really exist
If a user submits a malformed Director file, the status is correctly stored in
memcached, however `overlord-serve` will not submit the status since the project
does not really exist. In this case it is necessary to check if the status is
stored instead of checking for the existence of a project. A 404 not found error
is displayed for cases where no status is saved.

Unfortunately this removes support for using regular expressions when filtering
in `get-info`, as from now on VMs and projects are treated literally instead of
as a regular expression.