Avoid nodes with defective healthcheckers during autoscaling
The problem when we include nodes with healthcheckers in `unhealthy`
state is that the deployment will be successful, but nothing will
happen. The healthcheckers will not change their states in these
cases, the only solution is to fix the faulty service that cannot
be recovered using healthcheckers. But the motivation for this
change is not simply to ignore nodes with this use case, but to
deploy to a "good node" so that our service will start up
without problems.