Add `watch-commands` subcommand
In order to restart the whole group of processes without affecting
the running projects a new command has been created in order to
execute raw commands (or just system commands).

Now `watch-projects` and `watch-vm` take care of sorting the jobs
and creating the commands to be executed by `watch-commands`.
This also means that the code in `src/overlord/commands/watch.py`
has been migrated to a separate script (`libexec/create.py`).
This is another advantage of this design, as the code is loaded
each time the script is run, so there is no need to restart
processes for this purpose.

There are no longer `max_watch_projects` and `max_watch_vm`
parameters, it is now a single one called `max_watch_commands`.