Overwrite VM configuration after creating it
vm-bhyve does not respect some parameters (like uuid or networkX_mac)
that the user can specify from the template. This behavior makes
sense when vm-bhyve is used on a single machine but with multiple
virtual machines. However, since we use one virtual machine for each
jail, it might make sense for it to match the configuration we specify
from the template.